Fix breaking text on some complex subtitles

+++ src/ass/field.hpp
@@ -63,10 +63,10 @@ struct field_t {
 struct styles_spec_t {
     uint8_t alignment;
     int margin_v;
-    int explicit_y_pos;
+    float explicit_y_pos;
     bool is_drawing;
     explicit styles_spec_t();
-    styles_spec_t(uint8_t, int, int, bool);
+    styles_spec_t(uint8_t, int, float, bool);
 };
 
 /**
